functions/requirements.txt
# Azure Functions dependencies
azure-functions==1.17.0

# Database connectivity - pure Python driver (no C extensions)
pg8000==1.30.5

# Environment and utilities
python-dotenv==1.0.0

# This file must be in the ROOT of your function app
# Azure Functions looks for it here, not in functions/ subfolder
